KeyEvents :tada:

+++ UltralightNet/Enums/ULKeyEventModifiers.cs
@@ -1,7 +1,7 @@
 namespace UltralightNet
 {
 	/// <summary>An enumeration of the different keyboard modifiers.</summary>
-	public enum ULKeyEventModifiers: byte
+	public enum ULKeyEventModifiers: uint
 	{
 		/// <summary>Whether or not an ALT key is down</summary>
 		AltKey = 1 << 0,
